fix sync for dev_hdd0
no need to make the dev_hdd0 in the saves folder. once the config is pointed there, it gets made by the emulator.

going one level deeper on the check means we can make sure the files inside has been transferred.

the reason the script wasn't working was because on update / install of rpcs3 we were generating the dev_hdd0 folder.

then when we got to the save migration, checking for the folder we just created.