Readme Fix: name change mupen64plus FZ
This change is to fix a typo on what type of Mupen64plus because the FZ variant of mupen64plus is a android edition of the program which this isn’t installing. Don’t want to add any confusion if people take it literally. Especially since there is many variants of Mupen64plus like for example Mupen64plus-ae or mupen64plus-core. I redid PR because I didn’t follow dev guideline in readme and have no conflicts like previous.