Set up dual region Steam ROM Manager parsers (#465)
* Steam ROM Manager: Recognise games from platforms with multiple names

While both `genesis` and `megadrive` ROM directories are created when installing EmuDeck, only the `genesis` directory has an associated parser in Steam ROM Manager.

As a European, I used the `megadrive` directory as it's the name that comes to mind when I think of the console. So I was caught out when Steam ROM Manager didn't generate any entries for those Mega Drive games. This change fixes that issue.

Other platforms also have the same issue, so I've fixed the currently
known ones here.

Fixes:
- Sega Mega Drive / Genesis
- Sega CD / Sega Mega-CD
- TurboGrafx-16 / PC Engine
- TurboGrafx CD / PC Engine CD

* SRM: Fix globbing order to allow ROMs in subdirectories