helpers: Fix FlatPak installed check
The previous way of checking whether a package was installed or not was
incorrect in its assumption, that a package ID will only occur once in
the flatpak installed list.

I noticed this while trying to install PPSSPP, as FlatPak installs a
`org.ppsspp.PPSSPP` and `org.ppsspp.PPSSPP.Locale` package. The package
was installed correctly, but EmuDeck would register it as non-installed.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>