cloud: use more universal browser command that is not reliant on flatpak only
(1) This empties LD_PRELOAD in case anything is being preloaded into steam (such as libextest.so for wayland sessions). This way the preload does not carry into any browser we try to run.
(2) This adds a 'browsercommand' function that does these in order:

First, it tries the standard flatpak browserapp command.
If that fails, it then tries the the browserapp as a normal non-flatpak command
If that fails, it then uses xdg-settings to determine the default browser on the system then tries to launch that as a normal non-flatpak command.

This is a much more universal approach for non-flatpak reliant distros